refactor: 공통 메서드 AuthenticationInterceptor로 이동
