+++ src/main/java/nextstep/subway/auth/ui/interceptor/authentication/SessionAuthenticationInterceptor.java
@@ -1,59 +1,21 @@
 package nextstep.subway.auth.ui.interceptor.authentication;
 
 import nextstep.subway.auth.domain.Authentication;
-import nextstep.subway.auth.domain.AuthenticationToken;
 import nextstep.subway.auth.infrastructure.SecurityContext;
 import nextstep.subway.auth.ui.interceptor.authentication.converter.AuthenticationConverter;
 import nextstep.subway.member.application.CustomUserDetailsService;
-import nextstep.subway.member.domain.LoginMember;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
-import java.io.IOException;
 
 import static nextstep.subway.auth.infrastructure.SecurityContextHolder.SPRING_SECURITY_CONTEXT_KEY;
 
 public class SessionAuthenticationInterceptor extends AuthenticationInterceptor {
 
-    private CustomUserDetailsService userDetailsService;
-    private AuthenticationConverter authenticationConverter;
-
     public SessionAuthenticationInterceptor(CustomUserDetailsService userDetailsService,
                                             AuthenticationConverter authenticationConverter) {
-        this.userDetailsService = userDetailsService;
-        this.authenticationConverter = authenticationConverter;
-    }
-
-    @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws IOException {
-        AuthenticationToken token = authenticationConverter.convert(request);
-        Authentication authentication = authenticate(token);
-
-        if (authentication == null) {
-            throw new RuntimeException();
-        }
-
-        afterAuthentication(request, response, authentication);
-        return false;
-    }
-
-    public Authentication authenticate(AuthenticationToken token) {
-        String principal = token.getPrincipal();
-        LoginMember userDetails = userDetailsService.loadUserByUsername(principal);
-        checkAuthentication(userDetails, token);
-
-        return new Authentication(userDetails);
-    }
-
-    private void checkAuthentication(LoginMember userDetails, AuthenticationToken token) {
-        if (userDetails == null) {
-            throw new RuntimeException();
-        }
-
-        if (!userDetails.checkPassword(token.getCredentials())) {
-            throw new RuntimeException();
-        }
+        super(userDetailsService, authenticationConverter);
     }
 
     @Override
